added check for valid commentId, returns a rejected promise if the comment_id does not exist

+++ models/comments.model.js
@@ -1,6 +1,19 @@
 const db = require("../db/connection");
+const { checkExists } = require("../helpers");
 
 exports.removeCommentById = async ({ comment_id }) => {
+  const commentIdExists = await checkExists(
+    "comments",
+    "comment_id",
+    comment_id
+  );
+  if (!commentIdExists) {
+    return Promise.reject({
+      status: 404,
+      message: `Comment id "${comment_id}" not found in the comments table`,
+    });
+  }
+
   const queryString = `
         DELETE FROM comments
         WHERE comment_id = $1
